tests: Split out MockInventory from test-inventory
MockInventory will be reused for other test cases.
